[rename] 폴더 변경 및 import URL 수정
